Initialises React routes and creates first homepage component
Adds react-router-dom dependency to handle the routes for controller actions.
Creates Races component to display list of races on the homepage.
Creates App component which will handle all routes.
Amends packs/index.jsx to render the App component on load.
Adds axios dependency to get json data.

Races component fetches the data using axios and sets to the Race state array. The useEffect will run if the length of the array changes (after adding or deleting a record). The component creates an array of list items mapping the races array and returning an array of race names in a list component. The Races component returns an unorderded list of the races.

Fixes a spelling mistake in the db seed data.